replace the Classic/Crazy settings with a direct choice of maximum number of mines-per-tile
Move the mine counters and smiley face to the left of the playing field so that they don't make the window too wide when playing a Beginner game with multiple mines per tile.  Since this involves removing the flexible spacers sizeToContent() can now be used for resizing the window rather than setting innerWidth and innerHeight to 0 first.

Use the prefservice rather than persisted attributes to store settings.  Change the window id so that old persisted data is not used, because it seemed to be causing problems with window sizing