rewrite the game grid ui code using React.js
It's a bit of a mess for now, but it does work, and the code's already shorter and simpler than what was there before.

gridutil.jsx is based on code from netgame, though with no supporting for wrap-around, and with squares have eight adjacent tiles rather than four.  I might unify them at some point.